@@ -6,17 +6,17 @@ apply plugin: 'com.github.dcendents.android-maven'
 group='com.github.mundia416'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
 
 
 
     defaultConfig {
         minSdkVersion 19
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 
     }
 
@@ -32,18 +32,18 @@ android {
 
 configurations.all{
     resolutionStrategy{
-        force 'com.android.support:support-media-compat:27.1.1'
-        force 'com.android.support:support-v4:27.1.1'
+        force 'com.android.support:support-media-compat:28.0.0'
+        force 'com.android.support:support-v4:28.0.0'
     }
 }
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
-    implementation 'com.android.support:appcompat-v7:27.1.1'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    androidTestImplementation 'androidx.test:runner:1.1.0-beta02'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0-beta02'
     api 'com.google.android.gms:play-services-location:16.0.0'
     api 'com.google.android.gms:play-services-maps:16.0.0'
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
